+++ Changelog.md
@@ -15,6 +15,7 @@ OpenCore Changelog
 - Enforced the use of builtin picker when external fails
 - Fixed warnings for empty NVRAM variables (e.g. rtc-blacklist)
 - Added `ApplePanic` to store panic logs on ESP root
+- Fixed `ReconnectOnResChange` reconnecting even without res change
 
 #### v0.5.8
 - Fixed invalid CPU object reference in SSDT-PLUG